requirements.txt.txt
Flask
flask-bcrypt
flask-cors
pymongo
python-dotenv
openpyxl
dnspython
gunicorn